refactor: move rep view into rep file in views folder
